requirements.txt
opencv-python
pytesseract
numpy
Pillow
scikit-image
tensorflow
matplotlib
pytest
